+++ packages/terra-grid/tests/nightwatch/grid-spec.js
@@ -1,31 +1,21 @@
-/* eslint import/no-extraneous-dependencies: ["error", {"devDependencies": true}] */
 // eslint-disable-next-line import/no-extraneous-dependencies
-const screenshot = require('terra-toolkit').screenshot;
+const resizeTo = require('terra-toolkit/lib/nightwatch/responsive-helpers').resizeTo;
 
-module.exports = {
-  before: (browser, done) => {
-    browser.resizeWindow(browser.globals.width, browser.globals.height, done);
-  },
-
-  afterEach: (browser, done) => {
-    screenshot(browser, 'terra-grid', done);
-  },
-
+module.exports = resizeTo(['tiny', 'small', 'medium', 'large', 'huge', 'enormous'], {
   'Displays a default grid': (browser) => {
     browser
-      .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/grid-tests/default`)
+      .url(`${browser.launchUrl}/#/tests/grid-tests/default`)
       .assert.elementPresent('#grid');
   },
-
   'Displays a responsive grid': (browser) => {
     browser
-      .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/grid-tests/responsive`)
+      .url(`${browser.launchUrl}/#/tests/grid-tests/responsive`)
       .assert.elementPresent('#grid');
   },
 
   'Displays a nested grid': (browser) => {
     browser
-      .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/grid-tests/nested`)
+      .url(`${browser.launchUrl}/#/tests/grid-tests/nested`)
       .assert.elementPresent('#grid');
   },
-};
+});