+++ packages/terra-heading/tests/nightwatch/heading-spec.js
@@ -1,63 +1,54 @@
-/* eslint import/no-extraneous-dependencies: ["error", {"devDependencies": true}] */
 // eslint-disable-next-line import/no-extraneous-dependencies
-const screenshot = require('terra-toolkit').screenshot;
+const resizeTo = require('terra-toolkit/lib/nightwatch/responsive-helpers').resizeTo;
 
-module.exports = {
-  before: (browser, done) => {
-    browser.resizeWindow(browser.globals.width, browser.globals.height, done);
-  },
-
-  afterEach: (browser, done) => {
-    screenshot(browser, 'terra-heading', done);
-  },
-
+module.exports = resizeTo(['tiny', 'small', 'medium', 'large', 'huge', 'enormous'], {
   'Displays a default heading component': (browser) => {
-    browser.url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/heading-tests/default`);
+    browser.url(`${browser.launchUrl}/#/tests/heading-tests/default`);
     browser.expect.element('#heading-default').text.to.equal('Default');
     browser.expect.element('#heading-default').to.be.an('h1');
   },
 
   'Displays a heading component with a set color': (browser) => {
-    browser.url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/heading-tests/color`);
+    browser.url(`${browser.launchUrl}/#/tests/heading-tests/color`);
     browser.assert.cssProperty('#heading-color', 'color', 'rgba(48, 79, 254, 1)'); // Browser computes #304FFE to rgba(48, 79, 254, 1)
   },
 
   'Displays a heading component with a set size': (browser) => {
-    browser.url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/heading-tests/size`);
-    // Browser computes rem units as pixels when retrieved
-    browser.assert.cssProperty('#heading-size-huge', 'font-size', '32px');
-    browser.assert.cssProperty('#heading-size-large', 'font-size', '24px');
-    browser.assert.cssProperty('#heading-size-medium', 'font-size', '20px');
-    browser.assert.cssProperty('#heading-size-small', 'font-size', '18px');
-    browser.assert.cssProperty('#heading-size-tiny', 'font-size', '16px');
+    browser.url(`${browser.launchUrl}/#/tests/heading-tests/size`);
+    // Chrome browser computes to the hundredths, rem units to pixels, when retrieved
+    browser.assert.cssProperty('#heading-size-huge', 'font-size', '32.004px');
+    browser.assert.cssProperty('#heading-size-large', 'font-size', '23.996px');
+    browser.assert.cssProperty('#heading-size-medium', 'font-size', '20.006px');
+    browser.assert.cssProperty('#heading-size-small', 'font-size', '18.004px');
+    browser.assert.cssProperty('#heading-size-tiny', 'font-size', '16.002px');
     browser.assert.cssProperty('#heading-size-mini', 'font-size', '14px');
   },
 
   'Displays a heading component with set italics': (browser) => {
-    browser.url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/heading-tests/italics`);
+    browser.url(`${browser.launchUrl}/#/tests/heading-tests/italics`);
     browser.assert.cssProperty('#heading-italic', 'font-style', 'italic');
   },
 
   'Displays a heading component with set levels': (browser) => {
-    browser.url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/heading-tests/levels`);
-    // Browser computes rem units as pixels when retrieved
-    browser.assert.cssProperty('#heading-level-1', 'font-size', '32px');
-    browser.assert.cssProperty('#heading-level-2', 'font-size', '24px');
-    browser.assert.cssProperty('#heading-level-3', 'font-size', '20px');
-    browser.assert.cssProperty('#heading-level-4', 'font-size', '18px');
-    browser.assert.cssProperty('#heading-level-5', 'font-size', '16px');
+    browser.url(`${browser.launchUrl}/#/tests/heading-tests/levels`);
+    // Chrome browser computes to the hundredths, rem units to pixels, when retrieved
+    browser.assert.cssProperty('#heading-level-1', 'font-size', '32.004px');
+    browser.assert.cssProperty('#heading-level-2', 'font-size', '23.996px');
+    browser.assert.cssProperty('#heading-level-3', 'font-size', '20.006px');
+    browser.assert.cssProperty('#heading-level-4', 'font-size', '18.004px');
+    browser.assert.cssProperty('#heading-level-5', 'font-size', '16.002px');
     browser.assert.cssProperty('#heading-level-6', 'font-size', '14px');
   },
 
   'Displays a heading component with a set weight': (browser) => {
-    browser.url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/heading-tests/weight`);
+    browser.url(`${browser.launchUrl}/#/tests/heading-tests/weight`);
     browser.assert.cssProperty('#heading-weight-700', 'font-weight', 'bold'); // Browser computes 700 as bold
     browser.assert.cssProperty('#heading-weight-400', 'font-weight', 'normal'); // Browser computes 400 as normal
     browser.assert.cssProperty('#heading-weight-200', 'font-weight', '200');
   },
 
   'Displays a heading component with set hidden visibility': (browser) => {
-    browser.url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/heading-tests/visually-hidden`);
+    browser.url(`${browser.launchUrl}/#/tests/heading-tests/visually-hidden`);
     browser.assert.cssProperty('#heading-visually-hidden', 'border', '0px none rgb(28, 31, 33)'); // Browser computes 0 as 0px none rgb(28, 31, 33)
     browser.assert.cssProperty('#heading-visually-hidden', 'clip', 'rect(0px 0px 0px 0px)');
     browser.assert.cssProperty('#heading-visually-hidden', 'height', '1px');
@@ -69,7 +60,7 @@ module.exports = {
   },
 
   'Displays a heading component with all visual props set': (browser) => {
-    browser.url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/heading-tests/visual-props`);
+    browser.url(`${browser.launchUrl}/#/tests/heading-tests/visual-props`);
     browser.expect.element('#heading-visual-props').text.to.equal('All Visual Props Set');
   },
-};
+});