+++ packages/terra-site/src/examples/embedded-content-consumer/consumers/FillConsumer.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import EmbeddedContentConsumer from 'terra-embedded-content-consumer';
+
+const FillConsumer = () => (
+  <EmbeddedContentConsumer
+    src="#/tests/embedded-content-consumer-tests/fill-provider"
+  />
+);
+
+export default FillConsumer;