+++ packages/terra-grid/CHANGELOG.md
@@ -3,6 +3,13 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Update nightwatch tests and test scripts.
+
+3.9.0 - (August 31, 2017)
+------------------
+### Changed
+* Minor version bump
 
 3.8.0 - (August 15, 2017)
 ------------------